add refactor truss transfer
